Add files via upload
So far, we have added all vanilla classnames to the shops, and the majority of classnames from RHSUSF, RHSAFRF and RHSGREF. Also Greenmags and probably some over stuff based on our server's modlist. Trying to organize everything in the shops BY mod, so entire sections could be commented out in the config files depending on what mods are being used. Warning: still extremely WIP and pricing has not been set up yet, but take a look at the RHSUSF shops for a relative basis for what it might look like.